map rendering do not merge

+++ db/migrate/20190529211904_create_locations.rb
@@ -2,8 +2,9 @@ class CreateLocations < ActiveRecord::Migration[5.2]
   def change
     create_table :locations do |t|
       t.string :address
-      t.float :latitude
-      t.float :longitude
+      t.decimal :latitude
+      t.decimal :longitude
+
       t.timestamps
     end
   end
